remove all splash and info stuff... setup a common place for redirects so we dont break old links
