--- app/js/states/info/info.controller.js
@@ -1,14 +1,0 @@
-'use strict';
-
-var statesModule = require('../');
-
-/**
- * @ngInject
- */
-function InfoCtrl($scope, InsightsConfig, PermalinkService) {
-    $scope.gettingStartedLink = InsightsConfig.gettingStartedLink;
-
-    PermalinkService.scroll();
-}
-
-statesModule.controller('InfoCtrl', InfoCtrl);
