--- app/js/states/info/info.jade
@@ -1,103 +1,0 @@
-.info-page
-  .container.info-page-intro.text-center.band.band-hero
-    .row
-      .col-sm-12
-        insights-logo
-      .col-sm-8.col-sm-offset-2
-        insights-tagline
-
-  .band.band-white.band-narrow
-    .container
-      .row
-        .col-md-10.vcenter
-          section.section
-            h2.section-title(translate) Proactive systems management
-            p(translate) Red Hat Insights is a hosted service designed to help you proactively identify and resolve technical issues in Red Hat Enterprise Linux and Red Hat Cloud Infrastructure environments. Decreasing time spent discovering, researching, and resolving critical issues allows you to focus on driving high value, strategic initiatives.
-        .col-md-2.info-image.vcenter
-          img(src="static/images/Icon_RH_Misc_Lifestyle_RGB_Button.png", alt="Management")
-
-  .band.band-white.band-narrow
-    .container
-      .row
-        .col-md-10
-          section.section
-            h2.section-title(translate) Insights powered by Red Hat
-            p(translate)
-              strong The Insights engine
-              | is powered by the unparalleled expertise that comes from over 700 Red Hat Certified Engineers, over 30,000 documented technical solutions, and over 1,000,000 resolved issues. Based on Red Hat’s continuously expanding knowledge base, which includes expertise in third-party hardware and software integration, deterministic rules are constantly being created.
-
-  .band.band-white.band-narrow
-    .container
-      .row
-        .col-md-2.info-image.vcenter
-          img(src="static/images/Icon_RH_Software_Data_CMYK_Button.png", alt="Data Collection")
-        .col-md-10.vcenter
-          section.section
-            h2.section-title(translate) Lightweight data collection
-            p(translate) For analysis by the Insights engine, a minimal subset of information commonly provided in a standard support case is collected by the Insights client. This data is collected via secure protocols at defined intervals then analyzed and discarded at the next upload, or no longer than two weeks later.
-
-  .band.band-white.band-narrow
-    .container
-      .row
-        .col-md-10.vcenter
-          section.section
-            h2.section-title(translate) Clear, tailored, and actionable intelligence
-            p(translate) The Insights engine delivers results for review through the Insights UI in Red Hat Satellite and in the Red Hat Customer Portal. Through the intuitive Insights UI, configuration issues discovered during analysis are clearly displayed and prioritized by severity. Tailored remediation steps are presented to help you optimize workloads and confidently resolve critical issues and prevent future issues.
-        .col-md-2.info-image.vcenter
-          img(src="static/images/Icon_RH_Document_Blueprint_RGB_Button.png", alt="Actionable Intelligences")
-
-  .band.band-white.band-narrow#availability
-    .container
-      .row
-        .col-md-12
-          section.section
-            h2.section-title(translate) Availability
-            p(translate) Red Hat Insights is available for current Red Hat customers on Red Hat Enterprise Linux, delivered through either the Customer Portal or Red Hat Satellite. Your current subscription includes Red Hat Insights for up to 10 systems. To enroll additional systems, contact
-              a(href="http://www.redhat.com/en/about/contact/sales") sales
-              | .
-            .row
-              .col-md-6
-                table.table.table-cp.table-striped.table-bordered
-                  thead
-                    tr
-                      th(translate) Red Hat Insights supports
-                  tbody
-                    tr
-                      td(translate) Red Hat Enterprise Linux 6
-                    tr
-                      td(translate) Red Hat Enterprise Linux 7
-              .col-md-6
-                table.table.table-cp.table-striped.table-bordered
-                  thead
-                    tr
-                      th(translate) Delivered via
-                  tbody
-                    tr
-                      td(translate) Red Hat Customer Portal
-                    tr
-                      td(translate) Red Hat Satellite 5
-                    tr
-                      td(translate) Red Hat Satellite 6
-
-  .band.band-white.band-narrow
-    .container
-      .row
-        .col-md-12
-          section.section
-            h2.section-title(translate) Security comes standard
-            p(translate) Like all products and services from Red Hat, Red Hat Insights is developed with a focus on security and rigorous deployment controls.
-            p(translate)
-              a(ui-sref="info.security") Learn more
-              |  about how Red Hat Insights keeps your data secure while enabling you to proactively manage systems.
-
-  .band.band-white.band-narrow
-    .container
-      .row
-        .col-md-12
-          section.section
-            h2.section-title(translate) Additional resources
-            .row
-              .col-xs-12.col-sm-7.col-md-8.vcenter
-                a.btn.btn-app.cta-link.cta-link-sm.cta-link-white.info-links(href="static/misc/datasheet.pdf", target="_blank", translate) Datasheet
-                a.btn.btn-app.cta-link.cta-link-sm.cta-link-white.info-links(href='{{gettingStartedLink}}', translate) Getting started with Red Hat Insights
-                a.btn.btn-app.cta-link.cta-link-sm.cta-link-white.info-links(ui-sref='login', translate) Log in to Red Hat Insights
