--- app/js/states/splash/splash.jade
@@ -1,1 +1,0 @@
-p Loading&hellip;
