+++ app/styles/base.scss
@@ -127,7 +127,7 @@ $define-fa: true !default;
 @import 'components/webhooks';
 @import 'components/policies-modal';
 @import 'components/expand-all-button';
-@import 'components/splash/landing-page';
+
 @import 'components/prod-filters-link';
 @import 'components/prod-filters-modal';
 @import 'components/product-selector';
