--- app/styles/components/splash/_landing-page.scss
@@ -1,251 +1,0 @@
-.landing-page {
-  .band.ansible {
-    background: url("../static/images/bg_ansible.jpg") no-repeat;
-    background-size: cover;
-    color: #fff;
-
-    h2,
-    h3 {
-      color: #fff;
-    }
-
-    .logo {
-      @include rem('margin-bottom', $global-padding);
-      max-width: 240px;
-    }
-
-    a { color: #fff; }
-  }
-	video:hover {
-		cursor: pointer;
-	}
-	.top-page-wrap {
-		background: #000;
-
-		@include media-query(sm) {
-			background: url("../static/images/bg_landing-hero.jpg") no-repeat;
-			background-size: 100% auto;
-		}
-	}
-
-
-	// Landing page hero
-	.landing-page-top {
-		color: #fff;
-		text-align: center;
-		margin: 0;
-		background: url("../static/images/bg_landing-hero.jpg") no-repeat;
-		background-size: cover;
-
-		@include media-query(sm) {
-			background: none;
-		}
-
-		.btn {
-			@include font-size(20px);
-			@include rem('padding', 18px 32px);
-			background: #11c0f9;
-			text-transform: uppercase;
-        // margin-top: $global-margin;
-		}
-
-    .btn-seperator {
-      display: block;
-      color: $blue-dark;
-
-      @include media-query(sm) {
-        display: inline-block;
-        @include rem('padding', $global-padding);
-        color: #fff;
-      }
-    }
-  }
-
-	h1.headline { color: #fff; }
-
-	.band-white { background: #fff; }
-	.band-gray-light { background: #f7f7f7; }
-	.uppercase { text-transform: uppercase; }
-	.bold { font-weight: 700; }
-
-	.core-value {
-		@include media-query(sm) {
-			@include rem('padding', 0 30px);
-		}
-	}
-
-	.btn.btn-app:hover {
-		background: lighten(#11c0f9, 10);
-	}
-
-	.band-whispy {
-		background: url("../static/images/bg_whispy.jpg");
-		background-size: cover;
-	}
-
-	.band-black {
-		background: #252527;
-	}
-
-	.testimonial {
-		color: #fff;
-
-		.author {
-			@include rem('padding', 0 20px);
-			@include font-size(16px, 16px);
-			text-transform: uppercase;
-			//font-weight: 700;
-		}
-	}
-
-	.card {
-		.btn {
-			width: 100%;
-			display: block;
-			background: #cc0000;
-		}
-
-		p {
-			@include media-query(sm) {
-				min-height: 120px;
-			}
-			
-			@include media-query(md) {
-				min-height: 80px;
-			}
-
-			@include media-query(lg) {
-				min-height: 80px;
-			}
-		}
-
-		.card-title {
-			@include media-query(sm) {
-				min-height: 56px;
-			}
-
-			@include media-query(lg) {
-				min-height: auto;
-			}
-		}
-	}
-
-	.pre-footer {
-
-		.btn.btn-app {
-			@include font-size(20px);
-			@include rem('padding', 18px 32px);
-			background: #11c0f9;
-			text-transform: uppercase;
-			margin-top: $global-margin;
-		}
-	}
-
-	.core-values {
-		@include media-query(sm) {
-			@include rem('padding-top', 80px);
-			@include rem('padding-bottom', 80px);
-		}
-	}
-}
-
-.intro-video {
-  .videowrapper {
-    float: none;
-    clear: both;
-    width: 100%;
-    position: relative;
-    padding-bottom: 56.25%;
-    padding-top: 25px;
-    height: 0;
-  }
-
-  .videowrapper iframe {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-  }
-}
-
-.icon-center { text-align: center; }
-
-.icon-split {
-	position: relative;
-	width: 130px;
-	height: 130px;
-	margin: 0 auto;
-	display: inline-block;
-	@include rem('margin-bottom', $global-padding);
-	background: #0085cf;
-	color: #fff;
-	@include font-size(74px);
-	line-height: 1;
-	border-radius: 50%;
-	overflow: hidden;
-	text-align: center;
-
-	&:after {
-		content: "";
-		position: absolute;
-		top: 20%;
-		left: 30%;
-		width: 100%;
-		height: 130%;
-		background: lighten(#0085cf, 10);
-		transform: rotate(45deg);
-		z-index: 1;
-	}
-
-	i {
-		position: relative;
-		z-index: 2;
-		@extend %vertical-align;
-	}
-}
-
-.icon-split.icon-split-red {
-	background: #a30000;
-
-	&:after {
-		background: lighten(#a30000, 10);
-	}
-}
-
-.icon-split.icon-split-orange {
-	background: #ec7a08;
-
-	&:after {
-		background: lighten(#f4aa00, 10);
-	}
-}
-
-
-.landing-page .related-resources .card {
-  position: relative;
-
-  .btn {
-    @include media-query(md) {
-      @include rem('padding-left', 4px);
-      @include rem('padding-right', 4px);
-    }
-  }
-
-  .card-title, p { min-height: auto; }
-  p { @include media-query(md) { @include rem('padding-bottom', $global-margin); }}
-
-  @include media-query(md) {
-    @include rem('padding-bottom', 50px);
-  }
-
-  .full-width-button-wrapper {
-
-    @include media-query(md) {
-      @include rem('padding', 0 $global-margin);
-    }
-
-    &.bottom { @include rem('bottom', $global-margin); }
-
-  }
-}
