--- app/styles/components/splash/_splash-base.scss
@@ -1,129 +1,0 @@
-//colors
-$bg_blue: #4f5a69;
-$white: #fff;
-
-
-.splash {
-
-	.band-wispy {
-		@include rem('padding', $global-padding 0);
-		margin-bottom: 0;
-	}
-	.cta-link-red { margin-bottom: -1px; }
-
-	.cards { @include rem('padding', 30px 0); }
-
-	.img {
-		&.img-security { background-color: #c00; }
-		&.img-learnmore { background-color: #d1d4d3; }
-		&.img-gettingstarted { background-color: #08c0fc; }
-	}
-
-	.icon-split {
-		position: relative;
-		width: 130px;
-		height: 130px;
-		margin: 0 auto;
-		display: inline-block;
-		@include rem('margin-bottom', $global-padding);
-		background: #0085cf;
-		color: #fff;
-		@include font-size(74px);
-		line-height: 1;
-		border-radius: 50%;
-		overflow: hidden;
-
-		&:after {
-			content: "";
-			position: absolute;
-			top: 20%;
-			left: 30%;
-			width: 100%;
-			height: 130%;
-			background: lighten(#0085cf, 10);
-			transform: rotate(45deg);
-			z-index: 1;
-		}
-
-		i {
-			position: relative;
-			z-index: 2;
-			@extend %vertical-align;
-		}
-	}
-
-	.icon-split.icon-split-red {
-		background: #a30000;
-
-		&:after {
-			background: lighten(#a30000, 10);
-		}
-	}
-
-	.card {
-		text-align: center;
-	}
-}
-
-.splash .masthead {
-	text-align: center;
-	@include rem('margin-bottom', $global-margin);
-	@include rem('padding-bottom', $global-padding);
-
-	.logo-insights, .headline, .strapline { @include rem('margin-bottom', $global-margin); }
-
-	.headline {
-		@include font-size(36px);
-		line-height: 1;
-	}
-
-	.sep {
-		color: #fff;
-		@include font-size(24px);
-		font-weight: 100;
-		@include rem('margin', 5px);
-
-		@include media-query (xs) {
-			display: inline-block;
-			@include rem('margin', 18px 24px);
-		}
-	}
-	.strapline {
-		.light { @include font-size(21px); }
-		a {
-			color: #fff;
-			border-bottom: 1px solid;
-			&:hover { color: darken(#fff, 25); }
-		}
-	}
-
-	.btn { margin: 0; }
-}
-
-.splash .card {
-	@include font-size(20px);
-	@include rem('padding-bottom', $global-margin);
-	transition: .2s;
-
-	h2, h3, h4 {
-		@include font-size(24px);
-		line-height: 1;
-		text-transform: uppercase;
-	}
-
-	&:hover, &:focus {
-		background: #f2f2f2 !important;
-		border-color: #f2f2f2 !important;
-		color: #585858 !important;
-		box-shadow: inset 0 0 0 7px rgba(0, 0, 0, 0.1);
-
-		h1, h2, h3, h4, h5 {
-			color: #585858 !important;
-		}
-
-		.icon-split {
-//			box-shadow: 0 0 0 7px rgba(255, 255, 255, 1);
-			box-shadow: inset 0 0 0 7px rgba(0, 0, 0, 0.1);
-		}
-	}
-}
