+++ app/styles/insights.scss
@@ -16,10 +16,6 @@
 @import 'components/info/info-base';
 @import 'components/info/sticky-nav';
 
-// Splash Page
-//==================================================================*/
-@import 'components/splash/splash-base';
-
 // Reset portal styles
 @import 'components/portal/reset';
